refactor(core): rename ThreadLocalLinter to PooledLinter
That old name does not accurately describe what this rule does.
