+++ harper-core/src/linting/compound_nouns/mod.rs
@@ -32,7 +32,7 @@ merge_linters!(CompoundNouns => CompoundNounAfterDetAdj, CompoundNounBeforeAuxVe
 
 #[cfg(test)]
 mod tests {
-    use crate::linting::thread_local_linter::for_tests::create_test_pool;
+    use crate::linting::pooled_linter::for_tests::create_test_pool;
 
     create_test_pool!(CompoundNouns, CompoundNouns, CompoundNouns::default());
     use super::CompoundNouns;
