+++ harper-core/src/linting/its_contraction/mod.rs
@@ -14,8 +14,8 @@ merge_linters!(
 #[cfg(test)]
 mod tests {
     use super::ItsContraction;
+    use crate::linting::pooled_linter::for_tests::create_test_pool;
     use crate::linting::tests::{assert_lint_count, assert_no_lints, assert_suggestion_result};
-    use crate::linting::thread_local_linter::for_tests::create_test_pool;
 
     create_test_pool!(ItsContraction, ItsContraction, ItsContraction::default());
