+++ harper-core/src/linting/its_possessive.rs
@@ -152,7 +152,7 @@ impl ExprLinter for ItsPossessive {
 
 #[cfg(test)]
 mod tests {
-    use crate::linting::thread_local_linter::for_tests::create_test_pool;
+    use crate::linting::pooled_linter::for_tests::create_test_pool;
 
     create_test_pool!(ItsPossessive, ItsPossessive, ItsPossessive::default());
     use crate::linting::tests::{assert_lint_count, assert_no_lints, assert_suggestion_result};
