+++ harper-core/src/linting/modal_of.rs
@@ -131,8 +131,8 @@ impl ExprLinter for ModalOf {
 #[cfg(test)]
 mod tests {
     use super::ModalOf;
+    use crate::linting::pooled_linter::for_tests::create_test_pool;
     use crate::linting::tests::{assert_lint_count, assert_no_lints, assert_suggestion_result};
-    use crate::linting::thread_local_linter::for_tests::create_test_pool;
 
     create_test_pool!(ModalOf, ModalOf, ModalOf::default());
