+++ harper-core/src/linting/need_to_noun.rs
@@ -88,7 +88,7 @@ impl ExprLinter for NeedToNoun {
 
 #[cfg(test)]
 mod tests {
-    use crate::linting::thread_local_linter::for_tests::create_test_pool;
+    use crate::linting::pooled_linter::for_tests::create_test_pool;
 
     create_test_pool!(NeedToNoun, NeedToNoun, NeedToNoun::default());
     use super::NeedToNoun;
