+++ harper-core/src/linting/open_compounds.rs
@@ -151,8 +151,8 @@ fn get_compound_idx(toks: &[Token], src: &[char], compound: &str) -> Option<usiz
 #[cfg(test)]
 mod tests {
     use super::OpenCompounds;
+    use crate::linting::pooled_linter::for_tests::create_test_pool;
     use crate::linting::tests::{assert_lint_count, assert_suggestion_result};
-    use crate::linting::thread_local_linter::for_tests::create_test_pool;
 
     create_test_pool!(OpenCompounds, OpenCompounds, OpenCompounds::default());
