+++ harper-core/src/linting/plural_decades/four_digits.rs
@@ -90,7 +90,7 @@ pub fn match_to_lint_four_digits(
 
 #[cfg(test)]
 mod lints {
-    use crate::linting::thread_local_linter::for_tests::create_test_pool;
+    use crate::linting::pooled_linter::for_tests::create_test_pool;
 
     create_test_pool!(PluralDecades, PluralDecades, PluralDecades::default());
     use super::super::PluralDecades;
