+++ harper-core/src/linting/progressive_needs_be.rs
@@ -92,7 +92,7 @@ impl ExprLinter for ProgressiveNeedsBe {
 
 #[cfg(test)]
 mod tests {
-    use crate::linting::thread_local_linter::for_tests::create_test_pool;
+    use crate::linting::pooled_linter::for_tests::create_test_pool;
 
     create_test_pool!(
         ProgressiveNeedsBe,
