+++ harper-core/src/linting/pronoun_inflection_be.rs
@@ -169,7 +169,7 @@ impl ExprLinter for PronounInflectionBe {
 
 #[cfg(test)]
 mod tests {
-    use crate::linting::thread_local_linter::for_tests::create_test_pool;
+    use crate::linting::pooled_linter::for_tests::create_test_pool;
 
     create_test_pool!(
         PronounInflectionBe,
