+++ harper-core/src/linting/quite_quiet.rs
@@ -123,7 +123,7 @@ impl ExprLinter for QuiteQuiet {
 
 #[cfg(test)]
 mod tests {
-    use crate::linting::thread_local_linter::for_tests::create_test_pool;
+    use crate::linting::pooled_linter::for_tests::create_test_pool;
 
     create_test_pool!(QuiteQuiet, QuiteQuiet, QuiteQuiet::default());
     use super::QuiteQuiet;
