+++ harper-core/src/linting/redundant_self.rs
@@ -162,10 +162,10 @@ impl ExprLinter for RedundantSelf {
 
 #[cfg(test)]
 mod tests {
+    use crate::linting::pooled_linter::for_tests::create_test_pool;
     use crate::linting::tests::{
         assert_good_and_bad_suggestions, assert_no_lints, assert_suggestion_result,
     };
-    use crate::linting::thread_local_linter::for_tests::create_test_pool;
 
     use super::RedundantSelf;
