+++ harper-core/src/linting/simple_past_to_past_participle.rs
@@ -104,7 +104,7 @@ impl ExprLinter for SimplePastToPastParticiple {
 
 #[cfg(test)]
 mod tests {
-    use crate::linting::thread_local_linter::for_tests::create_test_pool;
+    use crate::linting::pooled_linter::for_tests::create_test_pool;
 
     create_test_pool!(
         SimplePastToPastParticiple,
