+++ harper-core/src/linting/then_than.rs
@@ -112,8 +112,8 @@ impl ExprLinter for ThenThan {
 #[cfg(test)]
 mod tests {
     use super::ThenThan;
+    use crate::linting::pooled_linter::for_tests::create_test_pool;
     use crate::linting::tests::{assert_lint_count, assert_suggestion_result};
-    use crate::linting::thread_local_linter::for_tests::create_test_pool;
 
     create_test_pool!(ThenThan, ThenThan, ThenThan::default());
