+++ harper-core/src/linting/web_scraping.rs
@@ -180,7 +180,7 @@ impl ExprLinter for WebScraping {
 
 #[cfg(test)]
 mod tests {
-    use crate::linting::thread_local_linter::for_tests::create_test_pool;
+    use crate::linting::pooled_linter::for_tests::create_test_pool;
 
     create_test_pool!(WebScraping, WebScraping, WebScraping::default());
     use crate::linting::tests::assert_suggestion_result;
